Further split tests according to situations
